request method check.

+++ api/app/Controller/CommentsController.php
@@ -16,6 +16,7 @@ class CommentsController extends AppController {
 	public $components = array('Api');
 
 	public function index() {
+		if (!$this->request->is('get')) throw new BadRequestException(__('It is a request that is not allowed.'));
 		$comments = $this->Comment->find('all');
 
 		if (!empty($comments)) {
@@ -26,6 +27,7 @@ class CommentsController extends AppController {
 	}
 
 	public function save() {
+		if (!$this->request->is('post')) throw new BadRequestException(__('It is a request that is not allowed.'));
 		if ($this->Comment->save($this->request->data)) {
 			$this->Api->success($this->Comment->find('first', array('conditions' => array('id' => $this->Comment->getLastInsertID()))));
 		} else {
@@ -34,6 +36,7 @@ class CommentsController extends AppController {
 	}
 
 	public function deleteId($id = null) {
+		if (!$this->request->is('delete')) throw new BadRequestException(__('It is a request that is not allowed.'));
 		if (!empty($id) && $this->Comment->delete($id)) {
 			$this->Api->success();
 		} else {
